+++ dandelion-upm/dandelion-upm-api/src/main/java/cn/icepear/dandelion/upm/api/domain/vo/MenuVO.java
@@ -18,7 +18,7 @@ public class MenuVO implements Serializable {
 	/**
 	 * 菜单ID
 	 */
-	private Integer menuId;
+	private Long menuId;
 	/**
 	 * 菜单名称
 	 */
@@ -27,7 +27,7 @@ public class MenuVO implements Serializable {
 	/**
 	 * 父菜单ID
 	 */
-	private Integer parentId;
+	private Long parentId;
 
 	private Integer menuType;
 
@@ -84,7 +84,7 @@ public class MenuVO implements Serializable {
 	@Override
 	public boolean equals(Object obj) {
 		if (obj instanceof MenuVO) {
-			Integer targetMenuId = ((MenuVO) obj).getMenuId();
+			Long targetMenuId = ((MenuVO) obj).getMenuId();
 			return menuId.equals(targetMenuId);
 		}
 		return super.equals(obj);
