+++ dandelion-upm/dandelion-upm-biz/src/main/java/cn/icepear/dandelion/upm/biz/service/SysUserRoleService.java
@@ -19,5 +19,5 @@ public interface SysUserRoleService extends IService<SysUserRole> {
 	 * @author 寻欢·李
 	 * @date 2017年12月7日 16:31:38
 	 */
-	Boolean removeRoleByUserId(Integer userId);
+	Boolean removeRoleByUserId(Long userId);
 }
